refactor: organize components directory
