+++ src/pages/Bmart/Home/index.tsx
@@ -4,13 +4,7 @@ import './style.scss'
 export type HomeProps = unknown
 
 const Home: React.FC<HomeProps> = (props) => {
-  return (
-    <>
-      <div className="home">
-        <h1 className="title">Hello World</h1>
-      </div>
-    </>
-  )
+  return <div className="home"></div>
 }
 
 export default Home
